main/requirements.txt
openai-whisper
numpy
torch
ffmpeg-python
boto3
